[localegen] adopt localecfg changes

+++ src/modules/localegen/main.py
@@ -3,6 +3,7 @@
 # === This file is part of Calamares - <http://github.com/calamares> ===
 #
 #   Copyright 2014, Philip Müller <philm@manjaro.org>
+#   Copyright 2014, Anke Boersma <demm@kaosx.us>
 #
 #   Calamares is free software: you can redistribute it and/or modify
 #   it under the terms of the GNU General Public License as published by
@@ -18,64 +19,54 @@
 #   along with Calamares. If not, see <http://www.gnu.org/licenses/>.
 
 import os
+import shutil
 
 import libcalamares
 
-def uncomment_locale_gen(locale, install_path):
-    """ Uncomment selected locale in /etc/locale.gen """
-
-    text = []
-    with open("%s/etc/locale.gen" % install_path, "r") as gen:
-        text = gen.readlines()
-
-    with open("%s/etc/locale.gen" % install_path, "w") as gen:
-        for line in text:
-            if locale in line and line[0] == "#":
-                # uncomment line
-                line = line[1:]
-            gen.write(line)
 
 def run():
-    """ Setup locale """
+    """ Create locale """
 
-    install_path = libcalamares.globalstorage.value( "rootMountPoint" )
+    us = '#en_US'
+    locale = libcalamares.globalstorage.value("lcLocale")
+    if not locale:
+        locale = 'en_US.UTF-8 UTF-8'
 
-    # Generate locales
-    keyboard_layout = libcalamares.globalstorage.value("keyboardLayout")
-    keyboard_variant = libcalamares.globalstorage.value("keyboardVariant")
-    # TODO: get locale
-    locale = 'en_US.UTF-8'
+    install_path = libcalamares.globalstorage.value("rootMountPoint")
 
-    uncomment_locale_gen(locale, install_path)
+    # restore backup if available
+    if os.path.exists('/etc/locale.gen.bak'):
+        shutil.copy2('%s/etc/locale.gen.bak' %
+                     (install_path), '%s/etc/locale.gen' % (install_path))
 
-    libcalamares.utils.chroot_call(['locale-gen'])
+    # run locale-gen if detected
+    if os.path.exists('/etc/locale.gen'):
+        text = []
+        with open("%s/etc/locale.gen" % install_path, "r") as gen:
+            text = gen.readlines()
+
+        # always enable en_US
+        with open("%s/etc/locale.gen" % install_path, "w") as gen:
+            for line in text:
+                if us in line and line[0] == "#":
+                    # uncomment line
+                    line = line[1:]
+                if locale in line and line[0] == "#":
+                    # uncomment line
+                    line = line[1:]
+                gen.write(line)
+
+        libcalamares.utils.chroot_call(['locale-gen'])
+        print('locale.gen done')
+
+    # Set etc/locale.conf
     locale_conf_path = os.path.join(install_path, "etc/locale.conf")
     with open(locale_conf_path, "w") as locale_conf:
-        locale_conf.write('LANG=%s\n' % locale)
+        locale_conf.write('LANG=%s\n' % locale.split(' ')[0])
 
+    # Set /etc/environment
     environment_path = os.path.join(install_path, "etc/environment")
     with open(environment_path, "w") as environment:
         environment.write('LANG=%s\n' % locale)
-
-    # Set /etc/vconsole.conf
-    vconsole_conf_path = os.path.join(install_path, "etc/vconsole.conf")
-    with open(vconsole_conf_path, "w") as vconsole_conf:
-        vconsole_conf.write('KEYMAP=%s\n' % keyboard_layout)
-
-    # TODO: check if this can be done in keyboard or already done
-    consolefh = open("%s/etc/keyboard.conf" % install_path, "r")
-    newconsolefh = open("%s/etc/keyboard.new" % install_path, "w")
-    for line in consolefh:
-        line = line.rstrip("\r\n")
-        if(line.startswith("XKBLAYOUT=")):
-            newconsolefh.write("XKBLAYOUT=\"%s\"\n" % keyboard_layout)
-        elif(line.startswith("XKBVARIANT=") and keyboard_variant != ''):
-            newconsolefh.write("XKBVARIANT=\"%s\"\n" % keyboard_variant)
-        else:
-            newconsolefh.write("%s\n" % line)
-    consolefh.close()
-    newconsolefh.close()
-    libcalamares.utils.chroot_call(['mv', '/etc/keyboard.conf', '/etc/keyboard.conf.old'])
-    libcalamares.utils.chroot_call(['mv', '/etc/keyboard.new', '/etc/keyboard.conf'])
 
     return None
